Deprecate ::VForVendetta
