--- doc/unreleased/default/v_for_vendetta.md
@@ -1,9 +1,0 @@
-# Faker::VForVendetta
-
-```ruby
-Faker::VForVendetta.character #=> "V"
-
-Faker::VForVendetta.speech #=> "Remember, remember, the Fifth of November, the Gunpowder Treason and Plot. I know of no reason why the Gunpowder Treason should ever be forgot..."
-
-Faker::VForVendetta.quote #=> "People should not be afraid of their governments. Governments should be afraid of their people."
-```
