+++ test/deprecate/test_deprecate_v_for_vendetta.rb
@@ -1,0 +1,21 @@
+# frozen_string_literal: true
+
+require_relative '../test_helper'
+
+class TestDeprecateVForVendetta < Test::Unit::TestCase
+  def setup
+    @tester = Faker::VForVendetta
+  end
+
+  def test_character
+    assert @tester.character.match(/\w+/)
+  end
+
+  def test_speech
+    assert @tester.speech.match(/\w+/)
+  end
+
+  def test_quote
+    assert @tester.quote.match(/\w+/)
+  end
+end
